Move TypeSettings to own file
